Move short flag into OptionsCli struct
I would like to keep all options from the CLI in the OptionsCli struct.
That way we can just move around that struct and we don't have to worry
about what specific value is sent to what function - we can just expect
to receive this struct and look for the value we're interested in.

All future flags and options should be added to this struct.